Complete migration to server version
