+++ communication/__init__.py
@@ -1,3 +1,3 @@
 from .base_interface import BaseInterface
-from .telegram.telegram_bot_server import TelegramInterface
+from .telegram.telegram_bot import TelegramInterface
 from .discord.discord_client import DiscordInterface
